+++ groups/bae/baenet/baenet_httpmessagetype.h
@@ -181,7 +181,9 @@ int baenet_HttpMessageType::maxSupportedBdexVersion()
 inline
 int baenet_HttpMessageType::fromString(Value *result, const bsl::string& string)
 {
-    return fromString(result, string.c_str(), string.length());
+    return fromString(result,
+                      string.c_str(),
+                      static_cast<int>(string.length()));
 }
 
 inline
