+++ groups/bae/baenet/baenet_httprequestheader.h
@@ -430,7 +430,7 @@ int baenet_HttpRequestHeader::accessFields(ACCESSOR& accessor) const
                             itf != d_extendedFields.endFields();
                           ++itf) {
         const char *fieldName       = itf->first.data();
-        const int   fieldNameLength = itf->first.length();
+        const int   fieldNameLength = static_cast<int>(itf->first.length());
 
         for (ConstFieldValueIterator itv  = itf->second.begin();
                                      itv != itf->second.end();
