+++ groups/bae/baenet/baenet_httpstatuscode.h
@@ -257,7 +257,9 @@ int baenet_HttpStatusCode::maxSupportedBdexVersion()
 inline
 int baenet_HttpStatusCode::fromString(Value *result, const bsl::string& string)
 {
-    return fromString(result, string.c_str(), string.length());
+    return fromString(result,
+                      string.c_str(),
+                      static_cast<int>(string.length()));
 }
 
 inline
